requirements.txt
pandas
sentence-transformers
torch
faiss-cpu
numpy